requirements.txt
python-dotenv==1.0.0
streamlit==1.26.0
langchain==0.0.291
openai==0.28.0
transformers==4.33.2
diffusers==0.21.1
accelerate
opencv-python==4.8.0.76
tiktoken==0.5.1
st-pages==0.4.4
streamlit_chat==0.1.1
streamlit_authenticator==0.2.3
streamlit_drawable_canvas==0.9.3
streamlit_extras
loguru==0.7.2
omegaconf
tritonclient[all]